requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mysqlclient==2.2.0
packaging==23.1
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
